+++ noether/catalogue/scientific/misc.py
@@ -1,12 +1,12 @@
 '''
-Essential scientific units.
+Miscelaneous or human scientific units.
 '''
 
 from noether.core import Unit
 
 from math import pi
 
-from ..prefixes import SI_all, SI_large, micro
+from ..prefixes import SI_all
 from ..fundamental import candela, kilogram, meter, kelvin, ampere
 from .si import radian, steradian
 from .si import joule, watt, pascal, cumec
@@ -59,26 +59,6 @@ metre_mercury = meter_mercury = Unit(
     info="defined by convention")
 mercury = meter_mercury / meter
 sverdrup = Unit(cumec * 1e6, "sverdrup", "Sv")
-
-# % Hartree atomic units
-# where hbar = a_0 = m_e = e = 1
-
-hartree = Unit(
-    joule(4.359_744_722_2071e-18, 8.5e-30),  # CODATA 2018
-    "hartree", "Eₕ")
-bohr_radius = a_0 = Unit(
-    meter(5.291_772_109_03e-11, 8e-21),  # CODATA 2018
-    "bohr_radius", "a_0")
-
-# % Small-scale
-
-atomic_mass_unit = amu = dalton = Unit(
-    1.660_539_066_60e-27 * kilogram,  # CODATA 2018
-    ['atomic_mass_unit', 'dalton'], ['amu', 'Da'], SI_large)
-
-micron = micro * meter
-angstrom = Unit(meter(1e-10), "angstrom", "Å")
-barn = Unit(1e-28 * meter**2, "barn", "b", SI_all)
 
 # % Photometric
 nit = Unit(candela / meter**2, 'nit', 'nit')